+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to this project will be documented in this file.
 - `contract` controller for contract management.
 - `tondev js wrap` generates `code` and `codeHash` fields in addition to `tvc`.
 - `tondev contract info` prints code hash.
-  
+- `tondev info` command prints a summary information from all controllers.
 - short aliases for controllers, commands and options.
   For example instead of using `tondev network list` 
   you can use `tondev n l` and even shorter `tondev nl`.